requirements.txt
confluent-kafka==2.3.0
supabase==2.3.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.0
numpy==1.26.0
pytest==7.4.0
apache-airflow==2.8.0
torch==2.1.0
scikit-learn==1.3.0
huggingface_hub==0.20.0
streamlit==1.29.0
plotly==5.18.0
